src/sign_bucket/requirements.txt
fastapi
pydantic
google-auth
google-auth-oauthlib
google-cloud-storage>=1.36.1
uvicorn
python-dotenv